security configuration was added

+++ src/main/java/ru/findFood/auth/repositories/RoleRepository.java
@@ -7,6 +7,7 @@ import ru.findFood.auth.entities.Role;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByTitle(String roleTitle);
+
 }
